+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Clearing filters after clicking "Previous"/"Next" buttons in Inventory tab on Browse form. UIIN-2131.
 * Also support `instance-storage` `9.0`, `holdings-storage` `6.0`, `item-storage 10.0`. Refs UIIN-2162.
 * Request with operator "==/string" doesn't return the exact match results when search for contributor name. UIIN-2157.
+* Fix clicking on the row by returning shelving order. UIIN-2171.
 
 ## [9.1.0](https://github.com/folio-org/ui-inventory/tree/v9.1.0) (2022-06-28)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.0.0...v9.1.0)